Refactor: move orchestration logic to standalone Orchestrator class
